Broke off some more tests
