+++ test/HiGlassComponentTest.js
@@ -3,7 +3,6 @@ import {
   configure,
   mount,
   // render,
-  ReactWrapper,
 } from 'enzyme';
 
 import Adapter from 'enzyme-adapter-react-16';
@@ -18,7 +17,6 @@ import HeatmapOptions from '../app/scripts/HeatmapOptions';
 
 // Utils
 import {
-  scalesCenterAndK,
   getTrackByUid,
   getTrackObjectFromHGC,
   getTiledPlot,
@@ -37,13 +35,10 @@ import {
   divergentTrackConfig,
   simpleCenterViewConfig,
   rectangleDomains,
-  threeViews,
   project1D,
-  onlyGPSB,
   chromInfoTrack,
   heatmapTrack,
   twoViewConfig,
-  oneViewConfig,
   oneTrackConfig,
   oneZoomedOutViewConf,
   // valueIntervalTrackViewConf,
@@ -1850,576 +1845,6 @@ describe('Simple HiGlassComponent', () => {
       );
 
       waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Ensures that the custom color map loads properly', (done) => {
-      // console.log('heatmap options:', HeatmapOptions);
-
-      hgc
-        .instance()
-        .tiledPlots.aa.handleConfigureTrack(
-          twoViewConfig.views[0].tracks.center[0].contents[0],
-          HeatmapOptions,
-        );
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-  });
-
-  describe('Close view tests', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'height:800px; width:800px');
-      div.setAttribute('id', 'single-view');
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: true }}
-          viewConfig={twoViewConfig}
-        />,
-        { attachTo: div },
-      );
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Ensures that when a view is closed, the PIXI graphics are removed', (done) => {
-      hgc.instance().handleCloseView('view2');
-
-      // console.log('hgc.instance:', hgc.instance().pixiStage.children);
-      // hgc.setState(hgc.instance().state);
-
-      // console.log('checking...', hgc.instance().pixiStage.children);
-      // since we removed one of the children, there should be only one left
-      expect(hgc.instance().pixiStage.children.length).toEqual(1);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-  });
-
-  describe('Three views and linking', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute(
-        'style',
-        'height:400px; width:800px;background-color: lightgreen',
-      );
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: true }}
-          viewConfig={threeViews}
-        />,
-        { attachTo: div },
-      );
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Links two views and moves to the side', (done) => {
-      hgc.instance().handleLocationLockChosen('aa', 'bb');
-      hgc.instance().handleZoomLockChosen('aa', 'bb');
-
-      hgc
-        .instance()
-        .tiledPlots.aa.trackRenderer.setCenter(
-          1799508622.8021536,
-          1801234331.7949603,
-          17952.610495328903,
-        );
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Checks to make sure that the two views have moved to the same place', () => {
-      const aaXScale = hgc.instance().xScales.aa;
-      const aaYScale = hgc.instance().yScales.aa;
-
-      const bbXScale = hgc.instance().xScales.bb;
-      const bbYScale = hgc.instance().yScales.bb;
-
-      const [aaCenterX, aaCenterY] = scalesCenterAndK(aaXScale, aaYScale);
-      const [bbCenterX, bbCenterY] = scalesCenterAndK(bbXScale, bbYScale);
-
-      expect(aaCenterX - bbCenterX).toBeLessThan(0.001);
-      expect(aaCenterY - bbCenterY).toBeLessThan(0.001);
-    });
-
-    it('Links the third view', (done) => {
-      hgc.instance().handleLocationYanked('cc', 'aa');
-      hgc.instance().handleZoomYanked('cc', 'aa');
-
-      hgc.instance().handleLocationLockChosen('bb', 'cc');
-      hgc.instance().handleZoomLockChosen('bb', 'cc');
-
-      hgc
-        .instance()
-        .tiledPlots.aa.trackRenderer.setCenter(
-          1799509622.8021536,
-          1801244331.7949603,
-          17952.610495328903,
-        );
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Makes sure that the third view moved', (done) => {
-      const aaXScale = hgc.instance().xScales.aa;
-      const aaYScale = hgc.instance().yScales.aa;
-
-      const ccXScale = hgc.instance().xScales.cc;
-      const ccYScale = hgc.instance().yScales.cc;
-
-      const [aaCenterX, aaCenterY] = scalesCenterAndK(aaXScale, aaYScale);
-      const [ccCenterX, ccCenterY] = scalesCenterAndK(ccXScale, ccYScale);
-
-      expect(aaCenterX - ccCenterX).toBeLessThan(0.001);
-      expect(aaCenterY - ccCenterY).toBeLessThan(0.001);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-  });
-
-  describe('Single view', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: false }}
-          viewConfig={oneViewConfig}
-        />,
-        { attachTo: div },
-      );
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('should load the initial config', (done) => {
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Changes the axis to inner right', (done) => {
-      const newOptions = {
-        axisPositionHorizontal: 'right',
-      };
-
-      hgc.instance().handleTrackOptionsChanged('aa', 'line1', newOptions);
-
-      const track = getTrackObjectFromHGC(hgc.instance(), 'aa', 'line1');
-      const { pAxis } = track.axis;
-
-      // we want the axis labels to be to the left of the end of the track
-      expect(pAxis.position.x).toBeGreaterThan(track.position[0]);
-      expect(pAxis.children[0].x).toBeLessThan(0);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Changes the axis to outside right', (done) => {
-      const newOptions = {
-        axisPositionHorizontal: 'outsideRight',
-      };
-
-      hgc.instance().handleTrackOptionsChanged('aa', 'line1', newOptions);
-
-      const track = getTrackObjectFromHGC(hgc.instance(), 'aa', 'line1');
-      const { pAxis } = track.axis;
-
-      // we want the axis labels to be to the left of the end of the track
-      expect(pAxis.position.x).toBeGreaterThan(track.position[0]);
-      expect(pAxis.children[0].x).toBeGreaterThan(0);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Changes the axis to outside left', (done) => {
-      const newOptions = {
-        axisPositionHorizontal: 'outsideLeft',
-      };
-
-      hgc.instance().handleTrackOptionsChanged('aa', 'line1', newOptions);
-
-      const track = getTrackObjectFromHGC(hgc.instance(), 'aa', 'line1');
-      const { pAxis } = track.axis;
-
-      // we want the axis labels to be to the left of the end of the track
-      expect(pAxis.position.x).toEqual(track.position[0]);
-      expect(pAxis.children[0].x).toBeLessThan(0);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Changes the axis to the left', (done) => {
-      const newOptions = {
-        axisPositionHorizontal: 'left',
-      };
-
-      hgc.instance().handleTrackOptionsChanged('aa', 'line1', newOptions);
-
-      const track = getTrackObjectFromHGC(hgc.instance(), 'aa', 'line1');
-      const { pAxis } = track.axis;
-
-      // we want the axis labels to be to the left of the end of the track
-      expect(pAxis.position.x).toEqual(track.position[0]);
-      expect(pAxis.children[0].x).toBeGreaterThan(0);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Changes the axis to the top', (done) => {
-      const newOptions = {
-        axisPositionHorizontal: null,
-        axisPositionVertical: 'top',
-      };
-
-      hgc.instance().handleTrackOptionsChanged('aa', 'vline1', newOptions);
-
-      const track = getTrackObjectFromHGC(hgc.instance(), 'aa', 'vline1')
-        .originalTrack;
-      const { pAxis } = track.axis;
-
-      // we want the axis labels to be to the left of the end of the track
-      expect(pAxis.position.x).toEqual(track.position[0]);
-      expect(pAxis.children[0].x).toBeGreaterThan(0);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Changes the axis to the outside top', (done) => {
-      const newOptions = {
-        axisPositionHorizontal: null,
-        axisPositionVertical: 'outsideTop',
-      };
-
-      hgc.instance().handleTrackOptionsChanged('aa', 'vline1', newOptions);
-
-      const track = getTrackObjectFromHGC(hgc.instance(), 'aa', 'vline1')
-        .originalTrack;
-      const { pAxis } = track.axis;
-
-      // we want the axis labels to be to the left of the end of the track
-      expect(pAxis.position.x).toEqual(track.position[0]);
-      expect(pAxis.children[0].x).toBeLessThan(0);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Changes the axis to the outside bottom', (done) => {
-      const newOptions = {
-        axisPositionHorizontal: null,
-        axisPositionVertical: 'outsideBottom',
-      };
-
-      hgc.instance().handleTrackOptionsChanged('aa', 'vline1', newOptions);
-
-      const track = getTrackObjectFromHGC(hgc.instance(), 'aa', 'vline1')
-        .originalTrack;
-      const { pAxis } = track.axis;
-
-      // we want the axis labels to be to the left of the end of the track
-      expect(pAxis.position.x).toBeGreaterThan(track.position[0]);
-      expect(pAxis.children[0].x).toBeGreaterThan(0);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Changes the axis to the bottom', (done) => {
-      const newOptions = {
-        axisPositionVertical: 'bottom',
-      };
-
-      hgc.instance().handleTrackOptionsChanged('aa', 'vline1', newOptions);
-
-      const track = getTrackObjectFromHGC(hgc.instance(), 'aa', 'vline1')
-        .originalTrack;
-      const { pAxis } = track.axis;
-
-      // we want the axis labels to be to the left of the end of the track
-      expect(pAxis.position.x).toBeGreaterThan(track.position[0]);
-      expect(pAxis.children[0].x).toBeLessThan(0);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('should have a bottom track of height 0', (done) => {
-      const height = hgc.instance().state.views.aa.tracks.bottom[0].height;
-      expect(height).toEqual(0);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-  });
-
-  describe('Value interval track tests', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'height:800px; width:800px');
-      div.setAttribute('id', 'single-view');
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: true }}
-          viewConfig={twoViewConfig}
-        />,
-        { attachTo: div },
-      );
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it("doesn't export maxWidth or filetype", () => {
-      const viewString = hgc.instance().getViewsAsString();
-
-      // expect(viewString.indexOf('1d-value-interval')).toBeGreaterThan(0);
-      expect(viewString.indexOf('maxWidth')).toBeLessThan(0);
-      expect(viewString.indexOf('filetype')).toBeLessThan(0);
-      expect(viewString.indexOf('binsPerDimension')).toBeLessThan(0);
-    });
-  });
-
-  describe('Starting with an existing genome position search box', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent options={{ bounded: false }} viewConfig={onlyGPSB} />,
-        { attachTo: div },
-      );
-
-      hgc.update();
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Makes the search box invisible', (done) => {
-      hgc.instance().handleTogglePositionSearchBox('aa');
-
-      waitForJsonComplete(done);
-    });
-
-    it('Makes the search box visible again', (done) => {
-      hgc.instance().handleTogglePositionSearchBox('aa');
-
-      waitForJsonComplete(done);
-    });
-
-    it('Searches for strings with spaces at the beginning', () => {
-      const gpsb = hgc.instance().genomePositionSearchBoxes.aa;
-
-      let [range1, range2] = gpsb.searchField.searchPosition(
-        '  chr1:1-1000 & chr1:2001-3000',
-      );
-
-      expect(range1[0]).toEqual(1);
-      expect(range1[1]).toEqual(1000);
-
-      expect(range2[0]).toEqual(2001);
-      expect(range2[1]).toEqual(3000);
-
-      [range1, range2] = gpsb.searchField.searchPosition(
-        'chr1:1-1000 & chr1:2001-3000',
-      );
-
-      expect(range1[0]).toEqual(1);
-      expect(range1[1]).toEqual(1000);
-    });
-
-    it('Ensures that hg38 is in the list of available assemblies', () => {
-      expect(
-        hgc
-          .instance()
-          .genomePositionSearchBoxes.aa.state.availableAssemblies.indexOf(
-            'hg38',
-          ),
-      ).toBeGreaterThanOrEqual(0);
-    });
-
-    it('Selects mm9', (done) => {
-      hgc.instance().genomePositionSearchBoxes.aa.handleAssemblySelect('mm9');
-
-      waitForJsonComplete(done);
-    });
-
-    it('Checks that mm9 was properly set and switches back to hg19', (done) => {
-      hgc.update();
-      const button = new ReactWrapper(
-        hgc.instance().genomePositionSearchBoxes.aa.assemblyPickButton,
-        true,
-      );
-      expect(button.props().title).toEqual('mm9');
-
-      hgc.instance().genomePositionSearchBoxes.aa.handleAssemblySelect('hg19');
-
-      waitForJsonComplete(done);
-    });
-
-    it('Checks that hg19 was properly', (done) => {
-      hgc.update();
-      const button = new ReactWrapper(
-        hgc.instance().genomePositionSearchBoxes.aa.assemblyPickButton,
-        true,
-      );
-      expect(button.props().title).toEqual('hg19');
-
-      waitForJsonComplete(done);
-    });
-  });
-
-  describe('Window resizing', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute(
-        'style',
-        'width:300px; height: 400px; background-color: lightgreen',
-      );
-      div.setAttribute('id', 'simple-hg-component');
-
-      const newViewConf = JSON.parse(JSON.stringify(project1D));
-
-      const center1 = JSON.parse(JSON.stringify(heatmapTrack));
-      const center2 = JSON.parse(JSON.stringify(heatmapTrack));
-
-      newViewConf.views[0].tracks.center = [center1];
-      newViewConf.views[1].tracks.center = [center2];
-
-      newViewConf.views[0].layout.h = 10;
-      newViewConf.views[1].layout.h = 10;
-
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: true }}
-          viewConfig={newViewConf}
-        />,
-        { attachTo: div },
-      );
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Sends a resize event to fit the current view into the window', (done) => {
-      const resizeEvent = new Event('resize');
-
-      window.dispatchEvent(resizeEvent);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Resize the view', (done) => {
-      div.setAttribute(
-        'style',
-        'width: 600px; height: 600px; background-color: lightgreen',
-      );
-      const resizeEvent = new Event('resize');
-
-      window.dispatchEvent(resizeEvent);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Expect the the chosen rowHeight to be less than 24', (done) => {
-      expect(hgc.instance().state.rowHeight).toBeLessThan(24);
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-  });
-
-  describe('Check for menu clashing in the center track ', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: false }}
-          viewConfig={simpleCenterViewConfig}
-        />,
-        { attachTo: div },
-      );
-
-      hgc.update();
-      waitForTilesLoaded(hgc.instance(), done);
-
-      // visual check that the heatmap track config menu is moved
-      // to the left
     });
   });
